+++ ios/RTNCastledNotifications/RTNCastledNotifications/RTNCastledNotifications.mm
@@ -21,12 +21,8 @@ RCT_EXTERN_METHOD(logCustomAppEvent:(NSString *)eventName eventParams:(NSDiction
 RCT_EXTERN_METHOD(setUserAttributes:(NSDictionary *)attributes);
 RCT_EXTERN_METHOD(onTokenFetch:(NSString *)token);
 RCT_EXTERN_METHOD(logout);
-RCT_EXTERN_METHOD(requestPushPermission);
-RCT_EXTERN_METHOD(userNotificationCenter:(NSDictionary *)userInfo);
-RCT_EXTERN_METHOD(willPresentNotification:(NSDictionary *)notificationInfo);
-RCT_EXTERN_METHOD(didReceiveRemoteNotificationInApplication:(NSDictionary *)notificationInfo completionHandler:(RCTResponseSenderBlock)completionHandler);
-RCT_EXTERN_METHOD(setNotificationCategoriesWithItems:(NSSet<UNNotificationCategory *> *)items)
-RCT_EXTERN_METHOD(setLaunchOptions:(NSDictionary<UIApplicationLaunchOptionsKey, id> *)launchOptions)
+RCT_EXTERN_METHOD(requestPushPermission: (RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(getPushPermission: (RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
 
 #ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule: